backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
langchain-core
requests
python-dotenv